src: fix Qt 6.11 + AppleClang 17 warnings (caught by macOS release build)
The local `make release` build (which mirrors CI's release.yml with
-Werror) flagged a handful of warnings that GCC + Qt 6.8 in CI doesn't
trip but AppleClang 17 + Qt 6.11 does:

- Application.cpp: drop unused `[this]` capture; QApplication::
  activeWindow() is static.

- ImageAdapter.cpp: QImage::mirrored() is deprecated in Qt 6.13;
  Qt 6.9 added flipped(Qt::Orientations). Use the latter under a
  QT_VERSION_CHECK guard until the minimum Qt moves to 6.9.

- AnnotationOverlay.cpp: float→qreal promotion in std::clamp call
  was implicit; make the cast explicit.

- MainWindow.cpp: drop unused `[this]` capture in the SAM/OCR
  download wireSignals lambda.

- ModelManagerDialog.cpp: qsizetype→int narrowing for setRowCount /
  loop bound; cast at the boundary.

- test_uat_foundations.cpp / test_uat_search_and_markup.cpp: same
  family of qsizetype→int and int→size_t conversion warnings in
  the UAT PDF-fixture builders.

None of these change runtime behaviour; they make the source clean
under stricter warning configs.

+++ src/ui/AnnotationOverlay.cpp
@@ -254,7 +254,8 @@ void AnnotationOverlay::paintEvent(QPaintEvent * /*event*/) {
             if (!a.pressures.empty() && a.pressures.size() == a.points.size()) {
                 const qreal base = a.style.strokeWidth > 0.0 ? a.style.strokeWidth : 1.5;
                 for (size_t i = 1; i < a.points.size(); ++i) {
-                    const qreal pr = std::clamp<qreal>(a.pressures[i], 0.0, 1.0);
+                    const qreal pr = std::clamp<qreal>(static_cast<qreal>(a.pressures[i]),
+                                                       0.0, 1.0);
                     // Same cubic curve as SignatureCanvas so a
                     // light touch is light and a heavy touch is
                     // confidently thick. base is the minimum.
